v2/requirements.txt
streamlit
faiss-cpu
pymupdf
sentence-transformers
numpy
pytest
requests
